Refactor to unify form creationg
